# docs/swagger.yaml
basePath: /api/v1
definitions:
  automation.RunResult:
    properties:
      contracts_near_expiry:
        type: integer
      monthly_report_generated:
        type: integer
      projects_suspended:
        type: integer
      recurring_charges_made:
        type: integer
      reminders_sent:
        type: integer
    type: object
...
  domain.User:
    properties:
      active:
        type: boolean
      balance:
        type: integer
      createdAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
      phoneNumber:
        type: string
      updatedAt:
        type: string
    type: object
...
  http.oidcAuthResponse:
    properties:
      auth_url:
        example: https://accounts.google.com/o/oauth2/auth?...
        type: string
      code_verifier:
        example: dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk
        type: string
      state:
        example: eyJhbGci...
        type: string
    type: object
  http.oidcCallbackRequest:
    properties:
      code:
        example: 4/0AX4XfWi...
        type: string
      code_verifier:
        example: dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk
        type: string
      state:
        example: eyJhbGci...
        type: string
    type: object
  http.tokenResponse:
    properties:
      access_token:
        example: eyJhbGci...
        type: string
      token_type:
        example: Bearer
        type: string
    type: object
  http.updateUserRequest:
    properties:
      firstName:
        example: João
        type: string
      lastName:
        example: Silva
        type: string
      phoneNumber:
        example: "11999999999"
        type: string
    type: object
  http.userPageResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/domain.User'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
  problem.BaseErr:
    properties:
      details:
        description: Details contém informações adicionais opcionais sobre o erro.
      error:
        description: Title é o título legível do tipo de problema.
        type: string
      errorDescription:
        description: ErrorDescription é uma explicação específica desta ocorrência.
        type: string
      instance:
        description: Instance é o URI da requisição que originou o erro.
        type: string
      statusCode:
        description: StatusCode é o código HTTP associado ao problema.
        type: integer
      timestamp:
        description: Timestamp indica quando o erro ocorreu (UTC, ISO 8601).
        type: string
      type:
        description: Type é um URI que identifica o tipo do problema (único por categoria
          de erro).
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: API modular de pagamentos com autenticação Bearer Token.
  title: AggiPay API
  version: 1.0.0
paths:
  /auth/google/callback:
    post:
      consumes:
      - application/json
      description: Troca o authorization code pelo JWT da aplicação usando PKCE. O
        cliente envia o code recebido do Google, o state e o code_verifier gerados
        no passo anterior.
      parameters:
      - description: Parâmetros do callback
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.oidcCallbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problem.BaseErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
      summary: Finalizar login com Google
      tags:
      - Auth
  /auth/google/login:
    get:
      description: Retorna a URL de autorização do Google e os parâmetros PKCE. O
        cliente deve armazenar state e code_verifier, depois redirecionar o usuário
        para auth_url.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.oidcAuthResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problem.BaseErr'
      summary: Iniciar login com Google
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Autentica com email e senha, retorna JWT de acesso
      parameters:
      - description: E-mail
        in: formData
        name: email
        required: true
        type: string
      - description: Senha
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problem.BaseErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
      summary: Login local
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Cria um novo usuário com email e senha. Usa criptografia Argon2ID
        com alto uso de processamento.
      parameters:
      - description: Primeiro nome
        in: formData
        name: firstName
        required: true
        type: string
      - description: Sobrenome
        in: formData
        name: lastName
        required: true
        type: string
      - description: E-mail
        in: formData
        name: email
        required: true
        type: string
      - description: Telefone
        in: formData
        name: phoneNumber
        type: string
      - description: Senha
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problem.BaseErr'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/problem.BaseErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/problem.BaseErr'
      summary: Registrar usuário
      tags:
      - Auth
  /auth/users:
    get:
      description: Retorna lista paginada de usuários
      parameters:
      - description: 'Página (padrão: 1)'
        in: query
        name: page
        type: integer
      - description: 'Tamanho da página (padrão: 10, máx: 100)'
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.userPageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Listar usuários
      tags:
      - Users
  /auth/users/{id}:
    delete:
      description: Realiza soft-delete de um usuário pelo ID
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Desativar usuário
      tags:
      - Users
    get:
      description: Retorna um usuário pelo ID
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Buscar usuário
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um usuário pelo ID
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      - description: Dados a atualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.updateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/problem.BaseErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Atualizar usuário
      tags:
      - Users
  /automation/run:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/automation.RunResult'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Executar automações financeiras
      tags:
      - Automation
  /automation/runs:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/problem.BaseErr'
      security:
      - BearerAuth: []
      summary: Histórico de automações
      tags:
      - Automation
...
securityDefinitions:
  BearerAuth:
    description: 'Informe: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"